Finish the project, returning two values in array representing the days
The logic of profit variable is changed. The second day must represent the sell day and the first day represents the buy day. Because you wouldn't call it a profit if it's higher "buy" and lower "sell". That and #abs is removed to find the negative returns.

Another logic is added to keep the sell days from going before the buy days.